+++ geotrek/core/static/core/topology_helper.js
@@ -65,12 +65,12 @@ Geotrek.TopologyHelper = (function() {
             }
         }
         else if (paths.length == 3 && polyline_start == polyline_end) {
-            var start_lls = polyline_start.getLatLngs()
-              , end_lls = polyline_end.getLatLngs();
+            var start_lls = polylines[0].getLatLngs()
+              , mid_lls = polylines[1].getLatLngs();
             cleanup = false;
             if (pk_start < pk_end) {
                 positions[0] = [pk_start, 0.0];
-                if (start_lls[0].equals(end_lls[0]))
+                if (start_lls[0].equals(mid_lls[0]))
                     positions[1] = [0.0, 1.0];
                 else
                     positions[1] = [1.0, 0.0];
@@ -78,7 +78,7 @@ Geotrek.TopologyHelper = (function() {
             }
             else {
                 positions[0] = [pk_start, 1.0];
-                if (start_lls[0].equals(end_lls[0]))
+                if (start_lls[0].equals(mid_lls[0]))
                     positions[1] = [1.0, 0.0];
                 else
                     positions[1] = [0.0, 1.0];
@@ -96,9 +96,11 @@ Geotrek.TopologyHelper = (function() {
             if (L.GeometryUtil.startsAtExtremity(polyline_start, polylines[1])) {
                 var next_lls = polylines[1].getLatLngs(),
                     next_end = next_lls[next_lls.length-1],
-                    share_end = first_end.equals(next_end);
+                    share_end = first_end.equals(next_end),
+                    two_paths_loop = first_end.equals(next_lls[0]);
                 if ((start_on_loop && pk_start > 0.5) ||
-                    (share_end && pk_start > 0.5 && pk_end > 0.5)) {
+                    (share_end && (pk_start + pk_end) >= 1) ||
+                    (two_paths_loop && (pk_start - pk_end) > 0)) {
                     /*
                      *       A
                      *    /--|===+    B
@@ -159,9 +161,11 @@ Geotrek.TopologyHelper = (function() {
             if (L.GeometryUtil.startsAtExtremity(polyline_end, polylines[polylines.length - 2])) {
                 var previous_lls = polylines[polylines.length - 2].getLatLngs(),
                     previous_end = previous_lls[previous_lls.length-1],
-                    share_end = last_end.equals(previous_end);
+                    share_end = last_end.equals(previous_end),
+                    two_paths_loop = last_end.equals(previous_lls[0]);
                 if ((end_on_loop && pk_end > 0.5) ||
-                    (share_end && pk_start > 0.5 && pk_end > 0.5)) {
+                    (share_end && (pk_start + pk_end) >= 1) ||
+                    (two_paths_loop && (pk_start - pk_end) <= 0)) {
                     /*
                      *              B
                      *     A    //==|-+