+++ geotrek/jstests/test.core.js
@@ -115,6 +115,11 @@ describe('Shortest path', function() {
 describe('Topology helper', function() {
 
     /*
+
+           40.0     +<---------+
+                  6 |          | 7
+           30.0     v----------+
+
            20.0     ^
                   3 |
                     |
@@ -146,7 +151,13 @@ describe('Topology helper', function() {
              properties: {pk: 4}},
             {type: "Feature",
              geometry: {type: "LineString", coordinates: [[10,-10], [0,-10]]},
-             properties: {pk: 5}}
+             properties: {pk: 5}},
+            {type: "Feature",
+             geometry: {type: "LineString", coordinates: [[0,40], [0,30]]},
+             properties: {pk: 6}},
+            {type: "Feature",
+             geometry: {type: "LineString", coordinates: [[0,30], [10,30], [10,40], [0, 40]]},
+             properties: {pk: 7}}
         ]
     }).addTo(map);
 
@@ -251,7 +262,7 @@ describe('Topology helper', function() {
 
     it('It should go through extremities only if it is the shortest way', function(done) {
         var topo = Geotrek.TopologyHelper.buildTopologyFromComputedPath(idToLayer,
-                        __inputData(L.latLng([10, 3]), L.latLng([2, 0]), [1, 2]));
+                        __inputData(L.latLng([5, 10]), L.latLng([2, 0]), [1, 2]));
 
         // One sub-topology
         assert.equal(topo.serialized.length, 1);
@@ -259,10 +270,48 @@ describe('Topology helper', function() {
         assert.deepEqual(topo.serialized[0], {
             offset: 0,
             positions: {
-                "0": [0.9010250453729739, 0.0],
+                "0": [0.5025512936152635, 0.0],
                 "1": [0.0, 0.20000000000000026]
             },
             paths: [1, 2]
+        });
+        done();
+    });
+
+
+    it('It should go through extremities, even if paths have opposite ways 2', function(done) {
+        var topo = Geotrek.TopologyHelper.buildTopologyFromComputedPath(idToLayer,
+                        __inputData(L.latLng([40, 3]), L.latLng([39, 0]), [7, 6]));
+
+        // One sub-topology
+        assert.equal(topo.serialized.length, 1);
+
+        assert.deepEqual(topo.serialized[0], {
+            offset: 0,
+            positions: {
+                "0": [0.9127652411579061, 1.0],
+                "1": [0.0, 0.10000000000000071]
+            },
+            paths: [7, 6]
+        });
+        done();
+    });
+
+
+    it('It should go through extremities only if it is the shortest way 2', function(done) {
+        var topo = Geotrek.TopologyHelper.buildTopologyFromComputedPath(idToLayer,
+                        __inputData(L.latLng([35, 10]), L.latLng([32, 0]), [7, 6]));
+
+        // One sub-topology
+        assert.equal(topo.serialized.length, 1);
+
+        assert.deepEqual(topo.serialized[0], {
+            offset: 0,
+            positions: {
+                "0": [0.5190218601242491, 0.0],
+                "1": [1, 0.8000000000000006]
+            },
+            paths: [7, 6]
         });
         done();
     });
@@ -308,7 +357,7 @@ describe('Topology helper', function() {
     });
 
 
-    it('It should work if loop with two paths', function(done) {
+    it('It should work if loop with two paths same way', function(done) {
         var topo = Geotrek.TopologyHelper.buildTopologyFromComputedPath(idToLayer,
                         __inputData(L.latLng([1, 0]), L.latLng([9, 0]), [2, 1, 2]));
 
@@ -323,6 +372,26 @@ describe('Topology helper', function() {
                 "2": [1, 0.8999999999999987],
             },
             paths: [2, 1, 2]
+        });
+        done();
+    });
+
+
+    it('It should work if loop with two paths opposite way', function(done) {
+        var topo = Geotrek.TopologyHelper.buildTopologyFromComputedPath(idToLayer,
+                        __inputData(L.latLng([31, 0]), L.latLng([39, 0]), [6, 7, 6]));
+
+        // One sub-topology
+        assert.equal(topo.serialized.length, 1);
+
+        assert.deepEqual(topo.serialized[0], {
+            offset: 0,
+            positions: {
+                "0": [0.8999999999999992, 1],
+                "1": [0, 1],
+                "2": [0, 0.10000000000000071],
+            },
+            paths: [6, 7, 6]
         });
         done();
     });